requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.6.1
future==1.0.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.45.0
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.3
numba==0.62.0
numpy==2.1.2
openai-whisper==20250625
pillow==11.0.0
regex==2025.9.18
requests==2.32.5
sympy==1.13.1
tiktoken==0.11.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.5.0